[refactor] 지도로더 훅 분리
